refactor: RacingSimulator 패키지 이동
